requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
crispy-bootstrap5==0.7
Django==4.2.10
django-crispy-forms==2.0
django-environ==0.11.2
djangorestframework==3.14.0
idna==3.4
Pillow==10.0.0
pytz==2023.3.post1
razorpay==1.4.1
requests==2.31.0
sqlparse==0.4.4
typing-extensions==4.7.1
tzdata==2023.3
urllib3==2.0.5
